- [DEV-491] Minor bug fixes and code improvements during code review.
git-svn-id: svn://svn.zabbix.com/branches/dev@15741 f4e7fd02-d210-4e34-a1e0-d4093fb66cc2

+++ dev-491/src/libs/zbxdbhigh/host.c
@@ -2647,7 +2647,7 @@ static int	DBcopy_template_items(zbx_uint64_t hostid, zbx_uint64_t templateid)
 						"interfaceid=%s"
 					" where itemid=" ZBX_FS_UI64 ";\n",
 					description_esc,
-					type,
+					(int)type,
 					row[4],		/* value_type */
 					row[5],		/* data_type */
 					row[6],		/* delay */
@@ -2706,7 +2706,7 @@ static int	DBcopy_template_items(zbx_uint64_t hostid, zbx_uint64_t templateid)
 					description_esc,
 					key_esc,
 					hostid,
-					type,
+					(int)type,
 					row[4],		/* value_type */
 					row[5],		/* data_type */
 					row[6],		/* delay */
